+++ HW0/Problem_1/.vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "files.associations": {
+        "utils.h": "c"
+    }
+}
