+++ HW0/Problem_1/.vscode/tasks.json
@@ -1,4 +1,5 @@
-{
+{   
+    "version": "2.0.0",
     "tasks": [
         {
             "type": "cppbuild",
@@ -6,9 +7,9 @@
             "command": "/usr/bin/gcc-9",
             "args": [
                 "-g",
-                "${file}",
                 "-o",
-                "src/* include/*"
+                "${fileDirname}/../build/${fileBasenameNoExtension}",
+                "${fileDirname}/*.c",
             ],
             "options": {
                 "cwd": "${workspaceFolder}"
@@ -16,12 +17,8 @@
             "problemMatcher": [
                 "$gcc"
             ],
-            "group": {
-                "kind": "build",
-                "isDefault": true
-            },
-            "detail": "Task generated by Debugger."
+            "group": "build",
+            "detail": "Task generated by Debugger.",
         }
-    ],
-    "version": "2.0.0"
+    ]
 }
