+++ HW0/Problem_1/src/main.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "bigint.h"
+#include "include/bigint.h"
 
 #define MAX_DEC 256
 
 int main()
 {
     
-    bigint* a;
+    bigint a = newnumc("34343");
     // newnumc("3434");
-
+    int i = 1;
     //scanf("%s %s", a, b);
