+++ HW0/Problem_1/src/utils.c
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "include/utils.h"
 
 //Choose the bigger one
 extm max(int* a, int* b)
